Add a SVG logo
This commit adds a SVG logo. I learned about the viewBox, and that you
set a coordinate system according to your needs. It defines the step size
you need to use to place units within the viewBox. Also, I learned that
you need to use preserveAspectRatio if the viewBox's container (div) has
different aspect ratio from the viewPort (like on mobile, which has a
different width than height). Also, a SVG will automatically fill the
container it is in, so this makes it responsive.