Add a button with hover and new font + colorway
Modeling the website after a theme I found. I learned how to include a
SVG file - inline <svg> in HTML allows way more control over fill, color
and size. Copy/paste the code from the SVG into HTML, it is the coords
of a path being drawn. THen you can place it inside a container and
CSS-modify the container, as the SVG will adapt to the container's size.
Added a hover effect, which was easy with .selector:hover, and I learned
that the font I am using has a special line height / ascent / other
characteristics, so I had to manually align it to be centered in the
button. I also learned how to change the child when I hover the parent,
it's parent:hover .childselected {style to apply}