Add hoverable icons and a content section
Renamed all the divs and filled them with content that fits the project.
Also made it all into flex-containers and made a basic layout.
Using transform:scale(float), and :hover, I made some nice hover effects
for the menu and the icons for the main sections. Now to add the big
logo and the slide-in / bar expanding animations.