Add a word-highlighting animation on scroll
Playing around with view() animations, which are not yet supported in
all browsers. It is a CSS-only alternative to using JS though. What It
does is check how much the container is in view. This works with keywords
like "cover" or "contain", or a fixed offset from the top of the
container.