Add a scrolling animation to the background image
The hero area presented a challenge. I wanted to have a bg image that I
could target directly with CSS selectors. What I had to do was making
a div in which I put both the text and the background-image. In order to
not have them next to each other, but stacked, I had to give them diffe-
rent z-indexes. This only works when you make the parent element
position:relative (so it is the nearest positioned parent element and
the position of surrounding elements is left intact) and the child gets
position absolute. You then apply z-indexes to them and position the
bg-image with background-size:cover.
For the scroll animation itself, I needed to use the scroller of the
root element on the animation-timeline (this is the body), as the img
has no scroller.