Fix the scrolling fade-in animation on hero
The issue was the timeline not being done correctly. I specify two
values: animation-timeline(end% start%). Start% is the % of the element
(or absolute offset, if in px) that specifies the distance between the
top of the viewport and the bottom of the element when the animation
starts. End% is the difference in element% between the bottom of the
element and the top of the viewport at animation end.

Also, currently applying it to entire hero container, it's likely I will
change this. Just wanted to get it working, prettify it later. Now on to
portfolio with hover effect.