Add portfolio section, fix units, change layout
Created a portfolio section with a nice transformX on an opacity:0 box
that therefore slides in. Also, changed the layout and updated the units
to rem to be more responsive. Will do the 62.5% font size on the .html
so the site adapts to user's font preference.
I also added the outline of a testimonial and footer section, which I
will now work on.