Add a autoplay / manual JS-based quote slider
This commit adds a quote slider, which can autoplay based on the
setInterval(func, delay) as well as get manually controlled based on
setting the class (classList.add) of the slide I want. This is done by
querySelectorAll, which gives an array, and then targetting by index.
I also remove the active class from the non-active ones so I can set the
opacity of the active one to 1 and to 0 on the others. All elements
are stacked on top of each other with position: relative / absolute,
and in essence I just change their opacity. Looks good (design TBD).